Mark options as required
The tests timeout on SSH connection if these options are not provided,
therefore mark them as such. Unfortunately due to a bug in oslo.config,
we can't do this in the "proper" way and must instead hack on things.
This hack can be removed once the bug is fixed upstream.

Change-Id: Ibc014f5c6f0315c993939e1cb9d13755da2af64f