Remove RefreshQuotaUsages
This feature was replaced upstream in OSP 12 (nova 16.0.0) [1] and the
test no longer works. The test would not make sense against the new
feature so simply remove it.

We are able to remove some utility methods that were only used by this
test. These are not used by anything else and should really be rewritten
if we introduce new tests that require these in the future.

[1] https://review.openstack.org/#/c/446243/

Change-Id: I1e6fa04503677db60d1b5e2c284ea03b834144fa