Convert get_hypervisor_ip to use hostnames
Instead of using ID to lookup hypervisor IPs, use hostnames. IDs are
different in undercloud and overcloud (a compute node's Ironic node
UUID in the undercloud isn't the same as its hypervisor UUID in the
overcloud). In addition, Nova uses both integer IDs (before 2.53) and
UUIDs (2.53 and later) for hypervisors. Hostnames are common to both
undercloud and overcloud, and don't change depending on the
microversion.

Change-Id: I34445e0f566dd48271740c357f74f632152776ed