Drop dependency on libvirt-python
Most existing tests are using our own VirshXMLClient (based on the
upstream tempest ssh client) and only the test_cpu_pinning.py tests
are using the libvirt client.

On one hand, using the libvirt-python client proves a bit more
complicated on a tripleo deployment, and it is taking too much effort
to get it to work on our infrared deployment.

On the other hand, all the other existing tests are using the ssh
client. If libvirt-python clients are to be introduced, it should be
inside of clients.py and it would be included in all the other tests.

(PS: Prior to this patch, the cpu pinning tests have not been run
successfully on our infrared deployment)

Change-Id: I2b65dca4fd26e9524b2b3ed7e182292eb5595599