Add max_compute_nodes option
This option indicates the number of (available) compute hosts in the
deployment, and will be used by tests that require all their instances
landing on the same host (ie, a single-compute deployment). The
default value is a childish large number to make sure that tests that
need a single node are skipped unless max_compute_nodes is set.

Change-Id: Id274baeb13ce87783e35daff74bf4cfe67554900