Add ServiceManager
Transform the ConfigClient into the ServiceManager. The ServiceManager
can modify a service's configuration and restart it. It can support
any service, as long as it's been added to the whitebox configuration.
This is done by having 'whitebox-<service-name>' sections in the
configration file. In addition, start using the crudini command
instead of ConfigParser. The latter isn't super easy to use in a
consistent way across Python versions.

Original-review: https://review.rdoproject.org/r/#/c/18755/
Change-Id: I38c847d938b3b83c0058f9d0033cfda9f98061bc