Readd vcpu variable for CPUPolicy classes
Commit 07c39615ae unified and refactored the create flavor method to a
single class within BaseWhiteboxComputeTest. Part of the commit removed
the vcpu class variable from BasePinningTest. CPUPolicyTest and
CPUThreadPolicyTest which inherit from BasePinningTest require the
variable to verify accurate pCPU pinning has been achieved in certain
testcases. Instead of reintroducing it back to BasePinningTest, this
patch adds it to the child classes that utilize the variable. In
addition CPUPolicyTest and CPUThreadPolicyTest originally relied on the
default parameter value of 2 vcpus from
base.BaseWhiteboxComputeTest.create_flavor() when creating flavors. If
the vcpu class variable needs to be adjusted for CPUThreadPolicyTest or
CPUPolicyTest it would cause potential failures since that vcpu count
was not explicitly passed to the create_flavor method, this patch
updates this to pass the class variable value now.

Change-Id: I8080c6319f98787eb663661d6aa0bd00bbea6d76