Check [whitebox]/hypervisors before using it
A devstack deployment will not set [whitebox]/hypervisors because the
host IPs are not known by the whitebox devstack plugin. Therefore,
instead of always using the [whitebox]/hypervisors option blindly,
this patch starts checking it for truthiness first. If false, we
fallback on using the hostname from the OS-EXT-SRV-ATTR:host field in
the show server response, or from the list hypervisors API. In a
devstack deployment, both of these should contain hostnames that are
resolvable and reachable from the machine running the tests.

This patch also reworks get_hypervisor_ip() to be cleaner.

Change-Id: I1a4430b6ee668bb34c847a63246a14b704185432