Skip test_rx_queue_size
test_rx_queue_size requires configuration that it is not setting
itself and relying on the deployment to set it. We can either refactor
it to set the configuration that it needs, or remove it altogether if
it adds no value. Until one of those two things happen, skip it.

Story: 2006820
Task: 37386
Change-Id: I22c73c023dcc6b9f4c5c67bd5bb8c24ef1edbb80