Don't use addition to combine dict values() into larger list
In Python 3, values() returns a dict view and not a list. Therefore,
if we attempts to add 2 values() together (like we would in Python 2)
to make a single list, we get:

TypeError: unsupported operand type(s) for +: 'dict_values' and
'dict_values'

This patch replaces addition with itertools.chain() to fix this.

Change-Id: Ie77e9fd417d4890e04672cedb62ee2f7972c8eda
Story: 2006826
Task: 37399