Fix getting server CPU pins
Commit b11da2e changed get_server_cpu_pinning() to return an int:set
dict. Turns out, all of the tests that used it set
cpu_policy=dedicated - thus pinning each vCPU to a single pCPU - and
expected an int:int dict. This commit reverts to the old behavior.
get_pinning_as_set(), which called get_server_cpu_pinning(), is
refactored to use the raw XML, and is moved down into
NUMALiveMigrationTest, as that's the only test class that uses it.

Story: 2006824
Task: 37397
Change-Id: I10fe6dcc49a5ca01a54d9ba95913f2aa6c0b1821