Wait after service restart
We need to make sure the service has actually started before
proceeding. Otherwise, in the case of nova-compute for example, we
might go on to boot a server, only for the service to restart in the
middle of the boot process. There is no straightforward and uniform
way to wait for a service to actually be running after a restart, so
we just sleep 5 seconds. This is an ugly hack and needs to be made
more robust in the future.

Story: 2006856
Task: 37449
Change-Id: I4fe8a14035148ed0a6b1c83cabb57e34ada756b2