Updates to run NUMALiveMigrationTest in tripleO
For downstream TripleO deployments, Undercloud is not setup to resolve
the IP addresses for compute/controller hostnames. To allow for an SSH
connection from the Undercloud to Overcloud databases,
NUMALiveMigrationTest.test_cpu_pinning was updated with a new call to
get_hypervisor_ip(). When setting up the ServiceManager() it will now
attempt to use the compute's IP address instead of the hostname. If no
IP information is found in [whitebox]/hypervisors, which is standard in
most devstack deployments, the hostname is utilized instead.  This is
the only instance where ServiceManager() utilized a hostname when
creating the service, all other ServiceManager() creations in
test_cpu_pinning.py use IP so no additional changes are necessary.

The current workaround for Nova bug 1836945 was only in place in
NUMALiveMigrationTest.test_cpu_pinning, added the same workaround to
NUMALiveMigrationTest.test_emulator_threads.

Iterator pagecounts in NUMALiveMigrationTest.test_hugepages was
exhausted via a skipException check before it could be used to calculate
the minimum free huge pages on a host. Iterator was 'reset' by creating
a new iterator of same var name.

Change-Id: I26855789d2da302c7356a71c42515ec4329cedd5