Increase service restart sleep from 5 to 15s
Current sleep time of 5 seconds after restarting a service is not always
sufficient enough for the service to become fully active in certain
environments. Increasing the sleep time from 5 seconds to 15. This
should be considered an interim fix, long term a more elegant solution
should be utilized for handling graceful service restarts.

Change-Id: I0b1786dd6cfa74d7ed315b0b3a92968a10a2c3e9