update requirements to support older releases
while python 2.7 support is nolonger required,
as this is a tempest plugin its possible desireable
to be able to run it on older releases.

This change caps the version of sshtunnel used on
python 2.7 to avoid a depency on cryptography 2.6
which is not avaliable on rhel 7 for downstream testing.

In the future we might want to consider if we drop
support for python 2.7 on the master branch and have a
seperate branch for 2.7 suport.

this change also removes py35 form the default tox envs
and update the tox config to skip missing interpreters and
ignore base python version conflict to allow pyXX envs
to work correctly.

Change-Id: I2726619df5752cd8478918346280244a8828c56c