Cap stestr at 2.6.0 for python27
The last stestr version to support Python 2.7 is 2.6.0. This has not
caused problems until recently, when we started seeing the py27 job
fail with the following error:

run.py discover: error: no such option: --list

This patch caps stestr to 2.6.0 when running with Python 2.7. It also
adds test-requirements.txt to the devstack job's irrelevant-files to
avoid wasting gate resources.

Change-Id: Ic05622530acdeb87d40e183e3e988042468c11d9