Follow up for I4660043fcccec6721e197c4d77390744e356c59e
This patch addresses nits from
I4660043fcccec6721e197c4d77390744e356c59e.

It also changes ctrlplane_address to ctlplane_address to bring it
inline with the exsting ctlplane_ssh_* options. It is NOT POLITE to do
so, but in practice, nothig is actually using the new ctrlplane
option in the wild yet, so it should be safe.

Change-Id: I9e06917a3ca6476fe4204f31e4121590140e90c9