Add vgpu guest creation testcase
Purpose of this commit is to test basic server creation while attaching
a vGPU device as documented for KVM systems here [1].

The commit adds a new whitebox configuration group ‘whitebox-hardware’.
This group currently contains one parameters: vgpu_vendor_id. It works
as both a skip check for the new testcase and as the specific gpu vendor
id to look for when analyzing the guest.

The new test method, test_create_vgpu_instance, was added to the file
test_vgpu.py in the test class VGPUTest. The test creates the necessary
flavor to launch a guest with an attached vgpu device and creates an
instance. It then ssh’s into the guest and validates the vgpu vendor id
is present when checking /sys/bus/pci/devices/<pci_address>/vendor.

A helper method get_pci_address() was lifted from nova.pci.utils [2].
This method is used to assemble the PCI components that are pulled from
the guest instance's device xml.

[1] https://docs.openstack.org/nova/train/admin/virtual-gpu.html
[2] https://github.com/openstack/nova/blob/master/nova/pci/utils.py

Change-Id: I736c0bf59c8c9c14aa8848d59077ef51071ffbbe