Add setuptools>=17.1 to test-requirements.txt
On RHEL 7 in downstream CI installation of whitebox fails due to the
presence of a version of setuptools that does not support the >=
operator for environment markers used in our requirement files, e.g.
python_version>='3.5'.

This change adds setuptools to the test requirements to allow
downstream CI jobs to automatically upgrade the setuptools version
before installing the whitebox plugin.

Change-Id: Idc00905cd1ae792d1bdde73214252106d094a6c7