New [whitebox-hardware]/cpu_topology option
This patch introduces a new config option that describes the CPU
topology of the compute host(s) in the test environment. No existing
tests are refactored (yet), but the plan is to get rid of automagic
introspection and replace it with declarative config, like the rest of
Tempest does. To that end, a couple of helpers are included in the new
hardware module.

Change-Id: I587b8cdcebef738bfa7b23267c54ccb90e4bc569