Change config_option to config_options
Previousy, the ServiceManager.config_option() context manager set a
single config option. This was wasteful when more than one config
option needed to be set for a single service, as that service ended up
being restarted as many times as there were config options to set.
This patch changes config_option() to config_options(), which now
accepts a list of (section, option, value) tuples. It'll set all of
them for the service, then restart it just once.

Change-Id: I2ce0687eb11b4d3ad6932a9993fa82e055c1ed09