Smarter Nova service management
This patch introduces a new NovaServiceManager class. This class uses
Nova's os-services API to make sure the service is actually up or
down, instead of dumbly sleep()ing.

Because the restart() method now calls stop and start in sequence, the
[whitebox-nova-compute]/restart_command config option is no longer
necessary, and has been removed. In its place, a new
[whitebox-nova-compute]/start_command option is needed.

Change-Id: I7b0c3635298b68981d9c460769a4d4ac9805357a