Add ensure-pip role
We've recently started getting Ansible pip error as follows:

ModuleNotFoundError: No module named 'pkg_resources'

This happens because we install crudini and iniparse from GitHub
source using Ansible's pip module in our Ansible roles, and Ansible's
pip module needs the system pip to be preconfigured (the exact
details are murkey - why this started happenig now is also a mystery).

This patch fixes that by adding the ensure-pip to our playbook. It
also renamed the plabook to be more consisten with other openstack
projects.

Change-Id: I3645d22afab1006e11cce22f42774ebea164998f