Use nested-virt nodepool label
To avoid being 100% dependant on the OpenEdge (formerly Fort Nebula)
nodepool provider, this patch starts using the nested-virt label. In
addition, because we can no longer expect SMT (aka Hyper-Threading)
to be present with the new nodepool label, the new configurable
[whitebox-hardware]/smt_hosts is introduced. It lists the hostnames of
compute hosts that support SMT.

We do have tests that require multi-NUMA and/or SMT. Those are moved
to a new non-voting job that runs on the multi-numa label provided by
OpenEdge.

Story: 2007395
Change-Id: If62d3a23044bf17f35a370e2f84fb459c166c9b2