Assert hugepages exist in domain XML
Add new checks to NUMALiveMigrationTest.test_hugepages to confirm
hugepages are present in the server XML and the page size is correct and
does not change after live migration. Added two new helper methods to
NUMALiveMigrationTest, _get_hugepage_xml_element and
_validate_hugepage_elements. The first method gets all hugepage XML
elements in the domain of a provided instance id. The second method
confirms there is only one hugepage xml element found on the instance
domain and that it is the correct page size. The testcase method
test_hugepages validates the hugepages element is present on both
instances and the configured pagesize matches the pagesize found on the
host. The test also rechecks server_b after it is live migrated to
confirm the element is still present and pagesize is accurate.

Fixed a problem in baremetal scenarios were determined ram value was a
float rather than an int. Explicitly setting it to int when providing it
to flavor creation.

Change-Id: I8fd32a68f2aea91145c0b444d1161f13ba1ae300