Normalize list of single dict for equality tests
Commit [1] normalized the json data for numa_topology to allow for
accurate comparison checks. Original logic would end the recursion when
a list was found. If the list had a dictionary, its list values would
not be sorted. This adds a check for dict objects in the list and
recurses on the object(s). It should be noted that when executing
v.sort(), there is a list of dictionaries greater than one, it results
in a TypeError exception since sorting against multiple dictionaries is
not supported by list method sort().

[1] https://review.opendev.org/#/c/705999/

Change-Id: I3abff34b84d14c854131e2631ad21911419fd20b