Introduce libvirt (un)mask)_command config option
On Ubunu Focal, libvirtd can be activated by other systemd units. To
avoid this unwanted activation, and only have libvirtd start when we
tell it to start, we need to mask it. This patch introduced the config
options to mask and unmask the libvirtd service. They will be used in
a susequent patch that starts using the Focal image to unskip some
tests.

Change-Id: Ib28811fd4f155cbc02f10b43d470f4cef142513b