move whitebox-multinode-multinuma-devstack to third party ci
This change removes whitebox-multinode-multinuma-devstack from
the check and gate pipeline since it will now be run by the
third party ci. whitebox-multinode-multinuma-devstack requires
multi numa host vms which are currently not avaiable in
the opendev ci.

This change makes whitebox-multinode-multinuma-devstack voting
by default. Since this will not be run in the first party
ci this will not block changes from mergeing on failure but
will prevent the third pary ci from reporting succes when the job
fails. This could be overriden externaly but voting is the default
and it is better to override that in the project section rather then
the job definition when required.

This change fixes the caulation of SMT_HOSTS to avoid undefined varible
errors in the ansible execution.

Change-Id: Iec3328933159b979f8d26c42247c88331be4caf5