Unskip test_cpu_model_extra_flags
test_cpu_model_extra_flags used to hardcode the cpu_model and
cpu_model_extra_flags that it expected. This is obviously very tied to
the deployment, and so was being skipped. This patch unskips the test
by adding configuration options that describe the cpu_model and
cpu_model_extra_flags that are set in the deployment.

Change-Id: Id32dcaab42ca94d98e65b869773717d1f9fa89e7