Remove restart_command config option
In real life, nothing actually uses that config option. We currently
support nova-compute and libvirt as services, and both only use
start_command and stop_command. Libvirt only uses those because it
needs to be stopped for a destructive test, and nova-compute only
uses those because it implements restart() as a sequence of stop/start
operations and waits for the service to appear down/up in the services
API at each step.

This patch removes the restart_command config option entirely, and
moves the stop()/start() implementation of restart() into the base
ServiceManager class.

If we ever actually need to restart a service, we can re-add this
config option in the future.

Change-Id: I14dddb5846b2330811c496aa7c134d020781edb8