Unskip test_rx_tx_queue_size
Previously, this test was skipped because it depended on a compute host
config that we did not set in CI. This patch adds a
[whitebox]/rx_queue_size config option to indicate what queue size is
configured on the compute hosts, gives it a sensible default in our
devstack plugin, makes test_rx_tx_queue_size use its value in the
test's assertion, and unskips the test. Woot. (The test itself also
needs a small fix in how it accesses the XML attribute).

Change-Id: I3cfe53dceb8801754b424ef7be71739a9b11c032