Added min_microversion to SRIOVNumaAffinity
Helper method create_test_server in the test methods for class
SRIOVNumaAffinity requires passing the 'host' parameter. This is only
supported for compute microversions 2.74 and greater. Adding class
variable min_microversion to to class SRIOVNumaAffinity to account for
this.

Also updated help msg for [whitebox-hardware]/physnet_numa_affinity to
be a more accurate description of the parameter.

Change-Id: I7fdc7f8fbfd6ee7329fdb4457d35a6ebdf5fe8d7