Update iniparse for py2.7 deployments
Current requirements when running with py2.7 on RHEL 7 systems results
in an installation error, see below. Setting the minimum acceptable
version for iniparse to 0.4 when using py2.7.

  Attempting uninstall: iniparse
    Found existing installation: iniparse 0.4

ERROR: Cannot uninstall 'iniparse'. It is a distutils installed project
and thus we cannot accurately determine which files belong to it which
would lead to only a partial uninstall.

Change-Id: I6109ae0dd67117875bd0b9011abb1c9d9c43ae14