Remove manager.EncryptionScenarioTest from TestQEMUVolumeEncryption
Introduced by Iceca320e5b93f7c64318648a94e775cc5d321c06 this test class
previously inherited from both compute api base and scenario manager
test classes.

When executed against an environment with Barbican deployed this would
lead to failures to delete secrets as the clients used by both of these
inherited classes differed, resulting in an unauthorised request to
delete a secret being made that failed.

To eliminate this the use of the scenario manager base class is removed
entirely and the logic required to create a volume type, encrypted
volume type and finally an encrypted volume is copied over verbatim. A
TODO is left in place to have this code refactored into a shared place
somewhere within Tempest in the future so both the API and Scenario
classes can have a unified approach to this.

Co-Authored-By: Lee Yarwood <lyarwood@redhat.com>
Change-Id: Id91887bb5d75c36c9a3c6fa4ea297c20b442fd4a