Add vgpu cold migration testcases
This commit introduces two new test methods, test_vgpu_cold_migration
and test_vgpu_revert_cold_migration. The test method takes it's cold
migration logical from tempest test [1] with some changes to highlight
specific areas to check for vGPU enabled instances.

Each method follows the same general procedure. The instance's current
compute host and a target compute host are determined. The vGPU resource
usage is calculated for both hosts and verified they meet expectations.
The instance is then cold migrated to the target host. For the
test_vgpu_revert_cold_migration the instance is then reverted back to
its source host when the initial migration is complete. At the
conclusion of the migration the instance's XML is checked to confirm the
vGPU XML element is still present and the vGPU's vendor id is present in
the guest. Lastly the compute hosts resource providers are rechecked to
confirm the vGPU resource usage is accurate.

The test class VGPUTest has been refactored so that it no longer creates
the vGPU instance and flavor at the start of every test. Instead during
resource_setup the flavor and instance is created and that same instance
is used for the entire suite of test methods.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/783140

[1] https://github.com/openstack/tempest/blob/stein-last/tempest/api/compute/admin/test_migrations.py#L136

Change-Id: I608822512dce81df18889cb310ce9e45f5ac049b