Start transition to declarative CPU pinning configs
Add two new jobs, one that runs CPU pinning tests with
[compute]cpu_dedicated_set, the other with [DEFAULT]vcpu_pin_set.
Start by converting the reboot live migration test to use the new way
of doing CPU pinning.

Because not all tests are converted at once (and the unconverted ones
are still changing the host CPU pinning configs on the fly), the new
jobs can only run the converted tests.

In the future, once all tests have been converted in subsequent
patches, we will merge the default job and the -cpupinning jobs,
leaving only the legacy job to run the CPU pinning tests that need
[DEFAULT]vcpu_pin_set.

In this patch, because we no longer always need the target_host
parameter to the live_migrate() helper, we can change its signature to
make target_host optional.

Change-Id: If0ef21f885eef880d55aed7ebed99828fe978d38