Remove NSM from SR-IOV Affinity Tests
With the transition away from the NovaServiceManager when working with
cpu_dedicated_set and cpu_shared_set, [1] SR-IOV NUMA affinity tests
needed to be updated. NSM no longer reconfigures the cpu_dedicated_set
for compute hosts involved in the test. Instead tests gather current
configuration of a target host and creates flavors based on the
parameters.

[1] https://opendev.org/openstack/whitebox-tempest-plugin/commit/cca33388c2ebce229fa3af70233769c3e2056d4d

Depends-on: https://review.opendev.org/c/openstack/whitebox-tempest-plugin/+/805731

Change-Id: I8340013dace577f415cf1b6d5e31f99ecd966964