Remove NSM from emul thread tests
A continuation of [1] in which we are trying to reduce the dependency on
NovaServiceManager when interacting with cpu_dedicated_set and
cpu_shared_set.

This commit introduces two new configuration options:
 * whitebox_hardware.dedicated_cpus_per_numa
 * whitebox_hardware.shared_cpus_per_numa

These options describe the number of pCPUs alloacted to either shared or
dedicated sets for a given compute host. We are making the assumption
that although the pCPUs allocated to their respective sets may be
different on each compute host, the number of pCPUs allocated are the
same.

Commit also removes NSM from three of the four base emulator thread
testcases and reduces some of the complexity originally present in the
tests. Since vcpu_pin_set and cpu_dedicated_set are no longer
directly modified in the test, their respective child tests have been
removed leaving just the parent to covering scenarios based on the
deployment.

[1] https://opendev.org/openstack/whitebox-tempest-plugin/commit/cca33388c2ebce229fa3af70233769c3e2056d4d

Change-Id: I57243d49de0d1715fecaa6496a92fc2c576c448a