Refactor ceph job definition
There's some weirdness going on with the ceph job once it gets
combined with barbican in the subsequent patch. No openstack services
are started on the subnode, and the only test ran in the job fails.
This appears to be pretty consistently reproducible.

Sidestep the problem entirely by switching our ceph job to inehrit
from the base devstack-plugin-ceph-multinode-tempest-py3 job. This
way, our whitebox-devstack-multinode job can get barbican in the
subsequent patch, and leave the ceph job barbican-less.

This does mean we have to duplicate some whitebox-specific
declarations in the ceph job, but it's not really worse than what we
had before, inheriting from the whitebox base job and duplicating the
*ceph job* declarations.

Also, since this job only runs storage tests, we don't need nested
KVM, and can use whatever nodeset we inherit from
devstack-plugin-ceph-multinode-tempest-py3.

[1] https://review.opendev.org/c/openstack/whitebox-tempest-plugin/+/779043

Change-Id: I6a576c0c54411bc83d482fb2db53cdae9033c19e