Remove admin client use for SR-IOV port creation
When applicable it is best to create/test resources with a non-admin
client. This commit reduces the need to use admin when creating ports.
In addition instead of using the 'host' parameter to force guest
scheduling with an admin client, use 'scheduler_hint' to direct guest
scheduling. 'scheduler_hint' does not require admin privileges, so
that reduces admin client use as well.

Change-Id: I1f0ba78fe56938182139042435ae6d62000d5c83