Take into account shared CPUs per NUMA
MixedCPUPolicyTest.test_shared_pinned_and_unpinned_guest was not taking
into account that only the shared PCPUs from the target NUMA node will
be allocated to the guest i.e. given the below topology:

[0]: 0,2,4,6,8,10,12
[1]: 1,3,5,7,9,11,13

PCPUs 0-3 are used as the shared set, with the remaining allocated
for dedicated set. Then scheduling a 2 vCPU guest with mixed policy
should would result in the guest either getting 0,2 or 1,3 only for its
shared PCPUs not 0-3.

This update checks that number of pCPUs allocated for shared is the
accurate length and a subset of the entire shared set.

Change-Id: I9bd375b563f67968ceae61d7db2ee86436dad771