Update grep cmd when checking for vgpu device
The original assumption that the XML pci bus address will be the same
within the guest when checking the guest for the vGPU device is
incorrect. Depending on the guest firmware the guest can renumber the
address with an address that is different from the XML.

Update no longer uses the pci address found in the XML and instead does
a wildcard grep search in /sys/bus/pci/devices/*/vendor for the vGPU
vendor ID. It also greps that output to confirm there is a valid PCI
address in the path.

Change-Id: Iacdc1947f4c0800aa49fd6d6ec9fae5b2da010d8