Add check for max disk attach
The commit [1] added a THT parameter that can set the maximum number of
disks that can be attached to a single guest. Refactored VirtioSCSIDisk
into ultimately three new classes.

 * VirtioSCSIBase - parent class hold standard setup and supporting
   methods for test cases
 * VirtioSCSIDiskMultiAttachment - repuprose of the test methods from
   VirtioSCSIDisk which are meant to test guest support of six or more
   disks
 * VirtioSCSIDiskRestrictAttachments - New test method that validates
   [1] is being enforced correctly.

Since [1] could be set to a value less than six, added skip checks for
VirtioSCSIDiskMultiAttachment if the max allowed disks that can be
attached are six or less.

[1] https://review.opendev.org/c/openstack/puppet-nova/+/708666/

Change-Id: I78f7f47ffa2644c71945bba2b5c1af29e0603363