Test coverage for stable compute UUID
This only covers that happy path of the feature. We agreed that the
error paths are hard to cover here and we rely on the nova functional
test instead.

There are two new config options:
* [compute_feature_enable]stable_compute_uuid_supported defaulted to
  False to skip the tests for old deployments not having the feature.
* [whitebox-nova-compute]state_path defaulted to /var/lib/nova, to allow
  running the tests in deployments that are using a non default nova
  state path configuration (like devstack)

The devstack plugin is modified to set [whitebox-nova-compute]state_path
according to the devstack nova configuration.

Change-Id: I520bd61b6ab479a36098586f7296822703a24fb0