Add vIOMMU device tests
Adds tests to verify [1] which allows operators to add vIOMMU devices to
their guests via flavor configuration. Tests do a basic smoke test that
creates a guest with the vIOMMU flavor requirements and confirms to
relevant XML element is present on the domain. Test also resizes a
standard guest to a flavor with vIOMMU, and confirms to XML is present.
Then revert and confirm the element is no longer present.

[1] https://blueprints.launchpad.net/nova/+spec/libvirt-viommu-device
Change-Id: I92c3538f985f0db0ef24ce2c49a797d3fcb69252