Add HyperV default enlightment tests
Tests update of adding libvirt default enlightenments for HyperV when
deploying a guest that uses image os_type 'windows'. [1] Test creates an
image and confirms the default attributes are present and in on state.

In addtion to cover regressions for [2], test that a guest with the
enlightenments can live migrate as well.

[1] https://review.opendev.org/c/openstack/nova/+/847641
[2] https://review.opendev.org/c/openstack/nova/+/904183

Change-Id: I9d8f3c4148949bfdd9fc956e271c18c4d76183b6