Deploy with CPU power management enabled
Until we fixed https://bugs.launchpad.net/nova/+bug/2056613 and
https://bugs.launchpad.net/nova/+bug/2056612, libvirt CPU power
management was broken with the `isolate` thread emualtor thread
policy, and with live migration. This was noticed more or less by
accident when the patch above this one triggered the right conditions
for our existing tests with incomplete coverage to break.

To prevent his from happening again, this patch turns on CPU power
management in our CI job, and modifies the test to assume that it's
turned on.

Change-Id: Ia48289ee4f909a33bfe6dac19eccff558b448da9