enable release jobs
This change adds the publish-to-pypi job template
for the whitebox-tempest-plugin.
This change also correct the syntax error in the
README.rst preventign the twine check command form passing.

this will enable releases to be done via commits to the release
repo via pypi https://pypi.org/project/whitebox-tempest-plugin/

a side effect of this process is the creation of git tags
and release tarballs for each tagged release.

Change-Id: I5cf0e3072c0502320866de1ce9771d61f1e0ec1b