add twine check and pre-commit
This change add a minimal pre-commit config to run
a local twine-check hook.

This hook emulates calls tools/twine-check.sh in a venv
which emulates the check made by the test-release-openstack
zuul job.

This can be run locally via "pre-commit run twine-check -a"
or automtically on commit if you install the pre-commit hooks

This change also add sphinx-lint to lint the readme and
catch common RST issues.

Change-Id: I726838714c591cc06d3bd54a47503bc6e659a0bb