Add a waiter when checking pci allocation
Intermittently the SR-IOV Live Migration Macvtap test would fail due to
the fact that the DB had the PCI allocation of the device as 2 after
migration instead of 1. This is due to the DB simply not updating in
time before the check is done.

Looking at Amit's working for evactuation [1] it makes sense to follow a
similar approach and have a waiter check the DB until it updates versus
doing a one time check.

[1] https://review.opendev.org/c/openstack/whitebox-tempest-plugin/+/915393/24/whitebox_tempest_plugin/api/compute/base.py#455

Change-Id: Icc152aa77cdced04cb67e15ab225c4bf754ca89e