Swtich to Ubuntu Noble image
Besides it being just a good thing to move to the latest LTS, this
also resolved an issue in different behaviour in libvirt versions for
watchdog devices. The q35 machine type always includes an implicit
watchdog device, but that device didn't start showing up in the XML
and having configurable actions until 9.1.0. On Jammy, the libvirt
version is 8.0.0, but in downstream RHOSO 18 CI on RHEL 9.4, the
version is 10.0.0. Moving to Noble unifies the libvirt versions, and
makes the watchdog testing less complicated.

Changes necessary to make this work:

* Start installing crudini with `package` instead of `pip` - new `pip`
  complains about the externally managed environment, and tells us to
  either use pipx to install in a venv, pass --break-system-packages,
  or install using system packages. Let's try the latter.

* Start running our local download job on changes to .zuul.yaml as
  well.

* Disable the br-ex-tcpdump service. It appears to only be useful for
  debugging, and it's crashing on the compute host with Noble. Just
  disable it for now.

Change-Id: I538f900953be2304baf8f63693699651fe763402